+++ ajax/itilfollowup.php
@@ -39,14 +39,9 @@
 
 use Glpi\Http\Response;
 
-/**
- * @var bool|null $AJAX_INCLUDE
- */
-global $AJAX_INCLUDE;
+/** @var $this \Glpi\Controller\LegacyFileLoadController */
+$this->setAjax();
 
-$AJAX_INCLUDE = 1;
-
-include('../inc/includes.php');
 header("Content-Type: application/json; charset=UTF-8");
 Html::header_nocache();