+++ front/cron.php
@@ -33,17 +33,20 @@
  * ---------------------------------------------------------------------
  */
 
+if (PHP_SAPI === 'cli') {
+    require(dirname(__DIR__) . '/vendor/autoload.php');
+
+    $kernel = new \Glpi\Kernel\Kernel();
+    $kernel->loadCommonGlobalConfig();
+}
+
 /**
  * @var array $CFG_GLPI
- * @var string|null $SECURITY_STRATEGY
  */
-global $CFG_GLPI,
-    $SECURITY_STRATEGY;
+global $CFG_GLPI;
 
 // Ensure current directory when run from crontab
 chdir(__DIR__);
-
-$SECURITY_STRATEGY = 'no_check'; // in GLPI mode, cronjob can also be triggered from public pages
 
 // Try detecting if we are running with the root user (Not available on Windows)
 if (function_exists('posix_geteuid') && posix_geteuid() === 0) {