+++ front/login.php
@@ -41,15 +41,8 @@ use Glpi\Application\View\TemplateRenderer;
 
 /**
  * @var array $CFG_GLPI
- * @var string|null $SECURITY_STRATEGY
  */
-global $CFG_GLPI,
-    $SECURITY_STRATEGY;
-
-$SECURITY_STRATEGY = 'no_check';
-
-include('../inc/includes.php');
-
+global $CFG_GLPI;
 
 if (!isset($_SESSION["glpicookietest"]) || ($_SESSION["glpicookietest"] != 'testcookie')) {
     if (!Session::canWriteSessionFiles()) {